Create anticipated_response_md
It is important to observe that Mark Z's conduct and questionable actions, as well as his moral compass and demonstrable lack of what can be termed "moral turpitude," were prominently exhibited during his appearance on the Joe Rogan podcast. He offered an apology to a vast audience, including members of Congress, in Washington, DC. My assertion is straightforward: while it may be true that he did not engage in any illegal activities, he is culpable for the fact that none of his omissions or apologies appeared genuine or were made proactively. In essence, his statements and apologies were issued only after being confronted or prompted to do so. This is just even more proof, " I will always be several steps in front of you and trying to claim you were the first here again ain't going to work here. "